requirements.txt
fastapi==0.109.2
pydantic==2.6.1
uvicorn==0.27.1
python-socketio==5.11.1
python-engineio==4.9.0
